Add blocks indexer, fix names of error codes files
Factored `default-header` from `nextname-header` and added it to
`cold` in order to restore the default behaviour after a cold restart.
This fixes a rare but possible bug, detected during the development of
`indexer`.

Made `need`, `needed`, `reneed` and `reneeded` deferred. Factored
`new-needed-word` from `needed`. These changes were needed to
implement `indexer`.

Added `string/`, useful for a check done by `indexer`.

Splitted and renamed the error codes files, in order to make sure they
are included in the right order.

Wrote `indexer`, a tool that indexes the library and makes `need`,
`needed`, `reneed` and `reneeded` much faster.

Added `cell/`, `d10*` and `bits` to the library.

Improved and updated the RNG benchmarks.