Rename library files
Many library files have been renamed. The goal is to build a logical
hierarchy, easy to search and extend, without using subdirectories, in
order to make it easy to select and link only the files needed by the
application. The task is not finished: Math, system, media, DOS and
other library files still have temporary names.