Fix `f=`, `f<>`, `f<`, `f<=`, `f>`, `f>=`
These words always returned true. The reason is their parameters are
compared as strings by the ROM calculator when called directly from
Forth. The origin of the problem is still unknow, but the details of the
debugging are noted in <src/lib/math.calculator.fsb>.

As a workaround, these six words have been rewritten with a new word
that executes the fp-calc-2 calculator command, which executes the
actual command, passed in the B register.