Add `user`
The code of `user` was inherited from fig-Forth. Beside, it had not been
adapted to DTC yet.  It has been fixed and renamed to `(user)`, as a
factor of a new version of `user`, which does not need the user data
pointer be specified. `?user` and `uallot` have been added too.