Change the stack notation for flags
The standard convention is "flag", though the shorter "f" is often.  So
far "wf" was used for well-formed flags, and "f" for zero/non-zero
flags, a non-standard distinction which sometimes is useful.

Now well-formed flags are represented by the normal "f", and
zero/non-zero flags, which are the exception, by "0f".