Move `warnings` control to the library
The check of redefinitions, done by `header,`, has been factored to a
new deferred word called `warn`, with defaults to `noop`, in order to
make it configurable during the development and save space for the
application.

The user variable `warnings` has been moved to the library,  and three
alternative behaviours for `warn` have been written.

Nevertheless, during the development of the system, `header,` keeps on
doing the check of redefinitions.