Rename `set` to `storer` and improve it
Though this usage of `set` was mentioned in Forth-79 and Forth-83, it's
seldom used, while `set` is a comus name to define a set of data, a more
useful feature which will be implemented.

`storer` is an obvious name for the old `set`. It has been improved with
double-cell operators, faster and shorter.