Fix `header,`: move `current-latest` to the kernel
A bug was introduced in `header,` when the standard `latest` replaced
the old fig-Forth `latest`, now called `current-latest`: new words were
linked to the latest word of the system, not to the latest word of the
compilation word list.

The solution is to move `current-latest` back to the kernel and use it
in `header,`.