Make the faster `<` definitive
The old version of `<`, adapted from DZX-Forth, was commented out
already. The version of `<` adapted from Z88 CamelForth is faster (0.96
vs 1.00), and only 9 bytes longer.

The source of the kernel has been revised in order to start indentifying
the operators that could be improved with a 16-bit comparison
instruction instead of calling the comparison routines adapted from
DZX-Forth.