Improve `need`
Sometimes `[needed]` and `[unneeded]` failed because `needed-word` was
overwritten in the circular string buffer.  The size of the buffer has
been doubled. the definitive solution will be to implement a string
stack in the system memory bank.

A simple test has been added for the debugging.  Documention has been
improved.