Improve compilation of 8-bit literals
`literal` is replaced with the new word `1literal` in `interpret-table`.
This causes 8-bit literals will be compiled by `cliteral` instead of
`literal`, what saves one byte of data space and is faster at run-time.
`cliteral` was moved from the library to the kernel, and `byte?` was
added in order to do the check.

These changes add 31 bytes to the kernel, but they will make the
programs smaller and faster.