Prepare low-level support for TR-DOS
The low-level code has been adapted from the G+DOS version.

The library TRD disk images are created, with the source on the sectors,
thanks to the implementation of `fsb2-trd`
(http://programandala.net/en.program.fsb2.html), but they can not be
tested, because they can not be open with the Fuse emulator, probably
because it examines sector 8 of track 0, where TR-DOS stores its own
meta data.  The solution is to use the track 0 from a formated disk
image, but a custom version of `fsb2-trd` is needed for that.