Compact the library
The code of several modules has been compacted and reorganized in order
to save blocks, because of the 636 KiB usable in TR-DOS disk images.

Further compression will be possible, because there are 1-block modules
that contain 1-line definitions, and they could be combined into one
single module with conditional compilation.