Rename disk images; add `blks` checker
All disk image files have been renamed after a shorter and clearer
format, easier to use with the SDL file interface of the Fuse emulator.
Now disks are numbered the same way in every DOS, with an ordinal number
instead the identifier of the drive it's supposed to be used with.

----
disks/
├── gplusdos
│   ├── disk0.mgt
│   ├── disk1_lib.mgt
│   ├── disk2_lib+games.mgt
│   ├── disk3_lib+benchmarks.mgt
│   ├── disk4_lib+tests.mgt
│   └── disk9_lib_without_dos.mgt
└── trdos
    ├── disk0.trd
    ├── disk1_lib.trd
    ├── disk2_lib+games.trd
    ├── disk3_lib+benchmarks.trd
    ├── disk4_lib+tests.trd
    └── disk9_lib_without_dos.trd
----

"disk9" is a special disk image used for debugging. It contains the core
library except the DOS modules. This means disk9 contains exactly the
same blocks in all DOS implementations. This is useful to check if disk
access works fine when a new DOS is implemented, comparing the output of
`blks` (a debug tool, temporarily included in the kernel) to a different
DOS.  The TR-DOS implementation has been checked this way. It works
fine, except the default drive can not be changed yet.