Add `lib-order`
So far the library must be one single disk.  `lib-order`,
`set-lib-order` and `get-lib-order` will make it possible to use several
library disks and configure the order in which they are searched, up to
the maximum allowed by the DOS. These words are not used by `need` yet,
so they are useless at the moment.