Try `indexer` and document it; improve README
The `indexer` did not work with the old 16 KiB "system bank" used to
store the definition headers, because the index occupied 12 KiB, what,
added to the 5 KiB used by the kernel definitions, filled the bank.
After implementing the new 64 KiB "far-memory" method, the indexer works
fine.

Its usage has been documented in the source and the README.