Move `?(` and `str=` to the kernel
Most library modules need `?(` for conditional compilation of
requirements (especially since `exit` is not used anymore to exit the
current block after loading the needed part), because `?\`, which is the
kernel from the first versions, is limited to one line.

The reason why `?(` was optional is that, in theory, the user
application could copy the source of the library modules as part of the
application, and arrange it such way that no dependency has to be
resolved during compilation. But that's unlikely, because using the
library as is, with its nested requirements, is much simpler. Therefore,
there's no point keeping `?(` in the library while it's needed by most
library modules.

Moving `?(` to the kernel forced moving `str=` as well.