Change filename extension of library files to "fs"
FSB is more a layout style than an independent format. Forth source can
be edited after this style in order to convert it to blocks with fsb2,
but in theory it could be interpreted as an ordinary text file by other
Forth systems as well, without a change.  Besides, the Vim modelines
make a specific filename extension unnecessary. Some changes were needed
in the fsb2 converter (http://programandala.net/en.program.fsb2.html).
Therefore it's simpler to use the usual filename extension "fs".