Make G+DOS don't change border during disk access
Changing the border color during disk access is a G+DOS specific
feature, which can be easily configured by the user storing a byte
bitmask into a G+DOS variable. By deactivating it by default, the basic
look and feel of the Forth system is identical on any DOS.  Information
on disk activity is provided by the status bar of the emulator.