Add first draft code to support block files
Block files can be supported easily on +3DOS (already done on
DZX-Forth); with some more effort on TR-DOS; and probably also on G+DOS,
after some low-level investigation.  But copying the library to a disk
image as a block file is not possible yet with the ordinary tools.

The only advantage of block files is having all files required to
compile a project (Solo Forth loader and binary, blocks, data files,
graphics, sounds, etc) in one single disk.

In practice, the problems to be solved are bigger than the possible
benefit.