Remove `save-display` tool; add `xy>r`, `r>xy`
`save-display` and `restore-display` were intended to be used before and
after calling a ROM routine that uses the display, which is a too
specific and seldom need that can be solved easily depending on each
case.

The parts `save-display` and `restore-display` that save and restore the
cursor have been extracted to `xy>r` and `r>xy`, which are useful
factors.

`full-display` and `nonfull-display` were just factors of `save-display`
and `restore-display`, without any other practical use, because the OS
variable they change has no effect in the Forth screen mode, it's used
only by BASIC.  Besides, changing that variable is needed only in code
words, before calling a ROM routine that may use the display.