Improve `to` and remove non-standard; add `cvalue`
The non-standard versions of `value`, `2value` and `cvalue` were aliases
of `constant`, `2constant` and `cconstant`. Their counterpart `to`,
`2to` and `cto` were aliases of `!>`, `2!>`. and `c!>`. This is faster
and simpler than the standard `to`, `value` and `2value`.

The duplicity between standard and non-standard `value`, `2value` and
`to` could be resolved with deferred words, but it's simpler to use
constants when speed is important, and modify them directly with `!>`,
`2!>` and `c!>`.

The standard `to` has been improved with an execution table and extended
to support also the non-standard `cvalue`.