Add the 64-cpl fonts of the 64#4 driver
Five fonts in the disk. They can be used with `mode-64o` (and the future
`mode-64s`).  One of them is identical to the current `owen-64cpl-font`
(formerly called `4x8-font`), already integrated in the library.
Eventually all of them will be included in the library.

`f64-test` shows all of the 64-cpl fonts.

Origin of the fonts:

	64#4 - 4x8 FONT DRIVER FOR 64 COLUMNS (c) 2007, 2011
	Original by Andrew Owen (657 bytes)
    Optimized by Crisis (602 bytes)
	Reimplemented by Einar Saukas (494 bytes)
	https://sites.google.com/site/zxgraph/home/einar-saukas/fonts
	http://www.worldofspectrum.org/infoseekid.cgi?id=0027130